requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bs4==0.0.1
captcha==0.2.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-appconf==1.0.2
django-contrib-comments==1.9.0
django-crispy-forms==1.7.2
django-guardian==1.4.9
django-imagekit==4.0.2
django-mdeditor==0.1.8
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-pagedown==1.0.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
django-taggit==0.23.0
django-threadedcomments==1.2
easy-thumbnails==2.5
html2text==2018.1.9
idna==2.7
jsonfield==2.0.2
lxml==4.3.0
Markdown==3.0.1
mistune==0.8.4
numpy==1.15.3
oauthlib==3.0.0
pilkit==2.0
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
python-slugify==2.0.1
python3-openid==3.1.0
pytz==2018.6
requests==2.21.0
requests-oauthlib==1.2.0
six==1.11.0
Unidecode==1.0.23
urllib3==1.24.1
